16 august 2023
Echo_botnrf52 is uploaded to icarus for receiving data from uart . Also can be uploaded to nrf52dk for sending data  to uart.
Broadacster is uploaded to nrf52dk, for advertising BLE beacons. It is not connectable but only scannable.
Observer is uploaded to gateway's nrf52 to scan beacons and forward the scanned data to uart.